fix DispatchEventAlt , tweak DoesFilterMatch w/ refr, WIP DumpEventHandlers
Fix DispatchEventAlt - it was storing the data of to-be-destroyed ArrayElements.
Instead, it now uses the data of the ScriptToken args directly.

DoesFilterMatch now assumes by default to compare the baseForm of an arg to its matching filter - this should be a more sensible default, for use with User-Defined Event dispatching via DispatchEventAlt.