Added array literals using the syntax from the 'in' expressions.
array myArr = [1,2,3,4]

These work with in expressions:
if (5 in myArr) {}

It is still more efficient to check directly against a literal if you can, as it compiles to faster code.

if (5 in [1,2,3,4]) {}

Adjusted for in loops to require the 'in' keyword. ':' no longer works.

Improved some error messages.

Adjusted the precedence of the pair operator :: - It will now bind tighter than * but weaker than unary ops.